+++ Source/ReleaseNotes.md
@@ -1,3 +1,4 @@
 * Core: Used new language features across the entire library
 * Core: Performance improvements
+* Server: Improved performance of retained messages when no event handler is attached (#2093, thanks to @zhaowgit)
 * Restored the strong name of the nugets